fix: CI integration tests ignored environment variable values (#31)
* fix: CI integration tests ignored environment variable values

The CI integration tests failed for a couple of reasons:
1. The spanner-testing instance is located in us-central1 which is no
   longer allow-listed for creating PG databases.
2. The staging endpoint currently returns an INVALID_ARGUMENT error for
   all CreateDatabase requests.
3. Changing the values for the above two in the environment variables
   for the CI script did not have any effect, as the Maven command
   executing the tests did not take these into account.

* fix: typo in mvn command + expected tables

* fix: remove table that is not yet available on prod

* fix: also change the row count

* fix: let the OS determine a random available port

The integration tests would assign a random port for PgAdapter to use if
no port was specified. This was however just a random number, and there
was no guarantee that that port was not already used by anything else on
the host machine. This could cause random build failures.

* fix: remove hardcoded options