refactor: replace JDBC with Connection API
Replaces the Cloud Spanner JDBC driver with the internal Connection API.
This reduces the number of times that we need to parse the SQL string
and translate between different types of parameters. It also enables us
to more easily make changes in the client library (Connection API) that
we can use for PgAdapter, as the Connection API is an internal API
without a strict contract on the interface.