fix: do not return Ready response for Copy messages
Copy messages should not respond with a Ready response, only with its
own specific Copy responses. Failure to do so can cause the calling
client to think that the operation has already finished, which again can
cause the entire copy operation to fail if the connection is closed
directly after receiving the Ready response.

Also the current Copy operation should not execute a rollback if
something fails, as it is operating in auto-commit mode. Trying to
rollback when in auto-commit mode will cause an error.