refactor: move pgx test to separate package (#81)
* build: only run Go tests in profile test-all

Disables running Go unit tests by default, so environments without a Go development
environment do not fail tests. Enabling the profile test-all will execute the Go unit
tests along with the Java tests. This profile is activated on the CI build.

* test: move pgx tests to a separate package

Moves the pgx tests to a separate package to make it easier to add other Go tests
in a different package from the pgx tests.

* test: add more Golang tests