feat: support connecting to different databases (#121)
Support connecting to different databases on an instance. This allows
for the use of the `\c` command in psql, as well as using a single
instance of PGAdapter to connect to multiple different databases using
other clients. The `-d` command line parameter is still supported, but
is now interpreted as follows:
1. If omitted, PGAdapter will connect to the database given in the
   connection request.
2. If included, PGAdapter will always connect to the database given in
   the -d command line parameter and ignore the database in the
   connection request.

Fixes b/194320320