fix: use assembly instead of fat jar
Customers should use an assembly with all dependencies in a separate
folder instead of a fat jar, as fat jars are prone to break if any of
the underlying dependencies use dynamic service loading and other forms
of dynamic class loading. The current fat jar for example breaks if it
is executed on a Google Cloud Engine VM using the default service
account of that VM, as the gRPC library will in that case attempt to
use DirectPath. That again requires certain classes to be dynamically
loaded, which breaks as not all classes and configuration files are
exactly where they are expected, after having been assembled into one
big jar.
A pre-built assembly does not have this problem, as all dependencies are
kept as separate jars including all their configuration files in their
original location. The pre-built Docker image already uses this
assembly. The downside of using an assembly instead of a fat jar, is
that it consists of more than one simple file. Customers must therefore
first untar the assembly and can then execute the pre-built jar with the
same command as before.