feat: automatically detect client that is connected
Automatically detects the client that is creating a connection to
PGAdapter based on the data that PGAdapter receives in the
StartupMessage. Not all clients can be uniquely identified this way, and
clients could easily pretend to be a client they are not, but this
feature is only used to automatically apply query replacements based on
the client that is connecting.

Fixes b/235348017