fix: send comments to the backend to support hints
Statement hints in PostgreSQL statements are placed inside comments.
Comments were removed from the SQL string before this was sent to the
backend. With this change, comments are preserved and sent to the
backend, so query hints are also picked up by the backend.

This change also fixes the limitations of the splitStatements parsing
that did not take dollar-quoted strings into account.

Fixes b/235426202