cleanup: remove custom log file for COPY operations
The COPY operation had a custom log file that was used to write any
errors that occurred during a COPY operation. This has now been replaced
by writing to the normal log file of PGAdapter, as there is no good
reason why COPY operations should log to a separate file.