feat: execute INFORMATION_SCHEMA in single-use transaction (#276)
* feat: execute INFORMATION_SCHEMA in single-use transaction

Queries referencing the INFORMATION_SCHEMA must use a single-use read-only
transaction. If such a query is executed during a (read/write) transaction, it
will return a specific error. This error is now detected by PGAdapter and the
query is retried in a separate single-use read-only transaction.

This behavior is needed for TypeORM.

* test: add more tests

* chore: add InternalAPI annotation

* fix: compilation failure