perf: treat 'flush and sync' as 'sync'
Some drivers unnecessarily send a Flush and a Sync message after an
extended query protocol sequence. This is unnecessary because Sync
already means that the pipeline should be flushed. It also causes an
inefficiency specifically in PGAdapter, as PGAdapter will choose a
transaction that is efficient for executing one query when it knows
that the implicit transaction only contains a single query. This is
not guaranteed if an extended query protocol sequence is ended by a
flush, as a flush could be followed by additional queries. PGAdapter
would therefore normally need to create an implicit read/write transaction
if it receives a Flush message after a query.

This change makes PGAdapter perform a look-ahead when it receives a
Flush message. If the next message is already available in the pipeline
and it is a Sync message, PGAdapter will treat the Flush message as if
it was a Sync message. This will ensure that a single query will use a
single-use read-only transaction even when it is followed by a flush
and sync message sequence.

The above behavior has been spotted in the NodeJS PostgreSQL driver.