feat: ignore named primary keys with correct name (#286)
CREATE TABLE statements that contain a named primary key constraint are
rejected by the backend, as the name of a primary key is fixed to 'PK_<table_name>'.
This change parses and removes named primary key constraint from CREATE TABLE
statements when these specify the same name as the generated name. This enables
the usage of frameworks that always generate CREATE TABLE statements with a named
primary key constraint, as long as the name that has been configured is equal to
the fixed name. The fact that the named constraint is stripped from the DDL
statement does not have any effect on the result of the statement, as the
generated name will be equal to the name specified in the statement.