feat: add ddl_transaction_mode and replace_pg_catalog_tables settings
Adds the ddl_transaction_mode and replace_pg_catalog_tables settings
from a purely server setting to one that can be set in session state.
This makes the features a lot more usable, as they can be set on a
per-session basis, and also set in for example the connection string.
This will for example enable the use of DDL transaction mode Explicit
for Liquibase.