fix: catch unknown types in RowDescription
Catch any errors when converting from a Spanner type to an OID to
prevent unknown types from making PGAdapter return an incomplete
response. An unknown type would now cause the RowDescriptionResponse to
fail halfway, which meant that half a message would be written to the
output buffer. This could then make a client hang, as it was waiting for
more data for the specific message.