docs: document max 50 parameters limit
PGAdapter can only describe statements with 50 query parameters or less.
This is caused by a limitation in the Cloud Spanner query engine that
does not allow a query that selects from a sub-query to contain more
than 50 columns, and this is what PGAdapter uses to fetch the parameter
types of a server side prepared statement.