perf: copy binary can write directly to a byte buffer (#385)
* perf: do not create a new parser for each column

Convert a Spanner value directly to the PostgreSQL equivalent without
creating a Parser object. This reduces the amount of memory that is
allocated and must be garbage collected for each row that is sent to
the PostgreSQL client.

* perf: write converted values directly to a byte buffer

Write the converted values directly to a byte buffer that can be
reused for multiple rows. This reduces the amount of objects that
are created and must be garbage collected for each result set that
is returned by PGAdapter.

* perf: copy binary can write directly to a byte buffer

Let binary copy operations write the converted data directly to a byte
buffer instead of creating a parser that puts it in a temporary byte
array, only to write it to the output stream later. This reduces the
number of objects that are created for each row that is copied.