fix: DDL batch errors halfway were not propagated
A DDL batch that executed at least one statement successfully and then
failed for a later statement would be reported as successful. This was
caused by an invalid extraction of the returned update counts from the
SpannerBatchUpdateException. The exception contains a full list of
update counts for all statements in the batch. The value is 1 for a
statement that was (successfully) executed and 0 for all other
statements. The extraction of the update counts assumed that the list
would only contain an entry for the statements that were actually
executed.

Fixes #443