cleanup: sample tests could fail silently and/or randomly
Writing the properties files for sample applications did not always work,
as the changes were not explicitly flushed to disk. This sometimes caused
an empty properties file to be written to disk.

Also fixes an issue in the Hibernate sample test that would cause the test
to get stuck if an error occurred, as the Hibernate SessionFactory was not
closed in that case.

Fixes #461