feat: support Connection API options in connection string (#542)
PostgreSQL supports sending session configuration parameters in the Startup
message. These can be embedded in the connection string of many clients, such
as JDBC. This worked for session configuration parameters that are maintained
in PGAdapter, but not for the very specific session configuration parameters
that are maintained by the Connection API, such as `spanner.autocommit_dml_mode`
or `spanner.read_only_staleness`. This PR adds support for those session
options in the StartupMessage.