feat: support more timestamp values for query parameters
PostgreSQL allows query parameters to be given in a wide range of input formats,
and the server will automatically convert this to the correct server value before
using these in the query. This adds support for lenient parsing of timestamp
values. These will be interpreted in the current timezone of the connection. This
means that if for example a timestamp value is given as just a date value, it
will be interpreted as midnight at the given date in the timezone of the current
connection.