fix: potential session leak when using prepared statements in autocommit mode
Creating multiple prepared statements in autocommit mode could cause a session
leak in PGAdapter, as the underlying result sets were not closed. This again
would hold on to the session, as there was no transaction that was being
closed after creating the prepared statement, which would otherwise have
returned the session to the pool.