feat: allow unsupported OIDs as param types
PGAdapter would refuse any parameter value that used an OID that was not
currently known to PGAdapter. This meant that if an application used for
example Oid.CHAR as the parameter type, PGAdapter would refuse it. This
change allows those parameters and automatically encodes these as untyped
parameter values using a string value when sending them to Cloud Spanner.
This is not guaranteed to work for all types, but it will work for any
type in Cloud Spanner that uses string encoding (meaning most of them).