feat: make table and function replacements client-aware
The current pg_catalog replacement strategy is fixed for all clients. This means
that if a specific client wants to use a pg_catalog table that is currently not
supported by Cloud Spanner, we need to add it for all clients. This can be tricky
if we cannot really support the table or function. This change makes the pg_catalog
implementation in PGAdapter client-aware. This also means that the initialization
of pg_catalog must use a lazy-initialized version of the WellKnownClient, as the
client may only be known once we get the first query. The detection of the client
based on the first query happens before the query is executed. This means that we
can safely assume that the WellKnownClient has been initialized when the backend
connection needs pg_catalog the first time.