fix: binary copy header should be included in first data message
The PostgreSQL wire protocol requires that the binary copy header is sent before any
data. It does however not specify whether it should be in a separate CopyData message,
or whether it should be included in the first actual CopyData message. PGAdapter used
to send the header in a separate message. This is accepted by most clients. PostgreSQL
sends the header as part of the first data message, and some clients (npgsql) assume
that this will always be the case.

This fix therefore changes the behavior of PGAdapter so it will always include the
header with the first actual CopyData message, instead of sending it in a separate
message. This also means that if the copy operation does not actually copy any data
at all, the header will be included in the trailer message. This is also consistent
with real PostgreSQL.