fix: show statements failed in pgx
SHOW variable_name statements failed in pgx, because pgx tries to describe all
statements (including SHOW). This failed as these statements do not have any
metadata, as they are not returned by the backend.

This fix also adds automatic detection of the pgx driver based on the first
Parse message that we receive. pgx uses a very specific name for the prepared
statements that it creates that can be used for (relatively) safe detection.