docs: emphasize that internal debugging is for internal use (#612)
* docs: emphasize that debug mode is for internal use

External users should normally not enable debug mode, but it has been observed
that some users still do. This changes the name of the command line argument to
emphasize that it is not an option that people should normally use. It also makes
sure that any existing use of the argument will silently be ignored.

* docs: emphasize that debug mode is for internal use