fix: do not execute analyze in a batch
The reWriteInserts=true JDBC option can cause the PG JDBC driver to
execute a batch that contains different SQL statements, where one of the
SQL statements in the middle of the batch must be described. The latter
happens if the statement contains at least one untyped parameter. The
JDBC driver sends all timestamp and date parameters as untyped, meaning
that statements with parameters of those types could cause this problem.
PGAdapter would then try to describe the statement as part of the batch,
which is not possible, as ExecuteBatchDml does not support PLAN mode.