fix: handle generic errors from ExecuteBatchDml
ExecuteBatchDml will normally throw a BatchUpdateException if something
goes wrong in a batch. It is however possible that the entire batch
fails, for example if the transaction has been aborted before the batch
is sent and retrying the transaction fails because of concurrent
modifications. Those errors would not be handled properly in PGAdapter
and could cause PGAdapter to return the generic exception
'Statement result cannot be retrieved before flush/sync' instead of the
actual underlying exception. This commit adds error handling and tests
for that possibility.