fix: close SpannerPool at shutdown
The server did not close the Spanner pool when it was gracefully shut down. This caused
gRPC channels to remain open when the server was shut down, which again could cause
applications to shut down more slowly, as they would wait for the threads that managed
those channels to shut down first.