feat: emulate pg_class and related tables (#766)
* feat: emulate pg_class and related tables

Adds common table expressions for pg_class and related tables. The OID of a
table is set as the fully qualified name of the table. This is not in line
with OIDs in PostgreSQL, as those are numeric, but most tools do not read
the OID anyways. Instead, the OID is used for joining the different pg_class
tables to get the relevant information.

This improves the compatibility for ORMs and tools that query pg_catalog for
information about the current objects in the schema, and is a requirement
for Ruby ActiveRecord. The feature can be turned off using the pg_setting
entry `spanner.emulate_pg_class_tables`.

* fix: make pg_type emulate-aware

* fix: add replacement for pg_type query

* fix: use '0' instead of '' to enable getting the value as an int

* test: add more tests

* chore: move tests to ITJdbcTest

* fix: move pg_class tests to a separate class

* test: add more tests

* fix: add ordering

* fix: order capital letters before uncapitalized

* test: add more tests