feat: support savepoints
Add support for emulated savepoints. Setting and releasing savepoints
work in all cases. Rolling back to a savepoint is not guaranteed to
work, as the underlying implementation will rollback the entire
transaction and retry up to where the savepoint was set. Note that
the retry will only be executed if the transaction is actually used
after the rollback.