feat: create local PostgreSQL copy
Adds a docker-compose.yml file and some scripts that will automatically
set up a local copy of a Cloud Spanner database in a PostgreSQL database.
The PostgreSQL database is created in a Docker container. The docker-compose
file also starts PGAdapter in a Docker container, and connects the real
PostgreSQL database to this PGAdapter instance, so that the Cloud Spanner
database is imported as a FOREIGN SCHEMA in the PostgreSQL database.