chore: update sample to use environment variables
Modified the sample slightly to use the same setup as the Go sample.
This means using environment variables for the Cloud Spanner database
that should be used, and starting PGAdapter without a default project,
instance and database. This makes the PGAdapter instance usable for
connections to different databases, which means that applications can
connect to multiple databases using a single PGAdapter instance.

It also shows how a single setup can easily be used for both test and
production.