docs: Cloud Run sidecar with Java app sample (#959)
* docs: Cloud Run sidecar with Java app sample

Adds a sample for running PGAdapter as a sidecar on Cloud Run together with an
application written in Java. The main application connects to PGAdapter using
the PostgreSQL JDBC driver and uses Unix domain sockets for the underlying
connection. The Unix domain sockets uses an in-memory volume.

* chore: update copyright year

* chore: update sample to use environment variables

Modified the sample slightly to use the same setup as the Go sample.
This means using environment variables for the Cloud Spanner database
that should be used, and starting PGAdapter without a default project,
instance and database. This makes the PGAdapter instance usable for
connections to different databases, which means that applications can
connect to multiple databases using a single PGAdapter instance.

It also shows how a single setup can easily be used for both test and
production.

* fix: add service replace step

* chore: remove duplicated line