feat: add Go wrapper for PGAdapter (#926)
* feat: native wrapper of PGAdapter in Go

Adds a native wrapper for starting PGAdapter using Go. This gives
Go developers a standard dependency that they can add to their
application and start/manage a PGAdapter instance in code.

* fix: module name

* fix: correct IsRunning check

* feat: add Go wrapper for PGAdapter

Create a Go wrapper for PGAdapter that can automatically download
and start PGAdapter from a Go application. This gives a similar
user experience as the in-process use of PGAdapter in Java and other
JVM-based languages.

* test: add GitHub test setup

* test: setup gcloud auth

* fix: check for relative path in zipped file