docs: use Options Builder for samples (#918)
* feat: create options builder

Add a builder for the OptionsMetadata class, to simplify building and
running PGAdapter as an in-process proxy. Instead of having to set
command line arguments in the Java application, users can just use a
normal Builder class to set options. This also significantly simplifies
setting structured options like SessionPoolOptions.

* docs: use Options Builder for samples

Use the new Options Builder for the Java in-process samples.

* fix: skip credentials check if Credentials instance has been set

* test: set default credentials for test

* fix: expect empty domain socket file for Windows

* fix: skip unix domain socket test on Windows

* feat: builder methods for channels and role

* docs: also update cloud-run sample