docs: use Go wrapper for gorm sample
Changes the `gorm` sample application to use the Go wrapper for
PGAdapter. This allows us to start PGAdapter directly from the sample
application and makes the ramp-up for running a simple sample a lot
lower. It also makes it possible to show that you can run the same
sample application on both Cloud Spanner PostgreSQL and open-source
PostgreSQL.