fix: replace fully-qualified column names
pg_catalog queries that include a fully-qualified column name in either
a join clause or a where clause would fail, as the pg_catalog table names
are replaced with a unqualified reference to a common table expression.

This change adds an additional step to replace fully qualified column
names of qualified table names that have been replaced, with column names
that only include the table name and not the schema name. That is:

`pg_catalog.pg_class.relname -> pg_class.relname`

Fixes #1129