fix: use array lower bound 1 by default (#1302)
* fix: use array lower bound 1 by default

PGAdapter would return arrays in binary format with lower bound
0 (zero) by default. This is not consistent with open-source
PostgreSQL, which uses 1 as the default lower bound, and also
not consistent with the text format, which does not have a field
for returning the lower bound. The lower bound is then assumed
to be 1. This changes the default lower bound for binary
encoded arrays to 1.

Fixes #1259

* test: add test data for integration test