fix: do not allocate statement name if it fails (#1252)
`PREPARE foo AS select` would fail, as `select` is not a valid SQL
statement, but the statement name `foo` would have been registered on
the connection. This meant that the user had to first DEALLOCATE the
statement again before rectifying the error. This change ensures that
the statement name is not registered on the connection if it failed.