build: Docker image with emulator (#1181)
* build: Docker image with emulator

* feat: build Docker image with PGAdapter+Emulator

Adds a Docker image to the build that contains both PGAdapter and the
Cloud Spanner Emulator. Running this Docker image will create a container
that starts both PGAdapter and the Emulator, and that automatically
configures PGAdapter to use the emulator in the Docker container.

The PGAdapter instance will accept connections for any valid database
name. If the database does not exist, it will automatically be created
on the emulator.

* fix: remove unnecessary changes

* feat: expose REST port for gcloud support

* docs: add documentation for pgadapter-emulator Docker image