sample: update Hibernate sample to use emulator (#1343)
Updates the Hibernate sample so it runs on the emulator. This makes it
both simpler for users to quickly run the sample, as well as making it
simpler to test the sample application, as all we need to do is run it
directly on the emulator.

Also updates the Hibernate dependency to the latest.