fix: limit the num bytes to write to length
Some Java 8 implementations return more bytes from wrapped CharSource
than there are. This would cause the converter to write 2 more bytes
than it should, which again would cause the next message in the buffer
to be wrong, as the last 2 bytes would become part of the length of
the next message. This could cause the stream to get blocked.