docs: remove samples for Cloud Run that embed PGAdapter in Docker image
Cloud Run originally did not support side-car containers. The only way to
deploy an application with PGAdapter on Cloud Run was therefore to include
PGAdapter in the Docker image. That is no longer necessary, and also not
recommended. So we should remove these samples and refer to the side-car
examples instead.