fix: correctly apply default client settings
Some auto-detected clients also set specific session state
variables. These variables were however not set if the client
was not auto-detected from the queries/parameters that the
client sends, but was set in the connection string
(e.g. Prisma). This fixes that so that the order of applying
session variables at startup are:
1. Set the values in the connction string.
2. Set the client based on the connection string (if any).
3. Apply the default params for the client, but do not
   overwrite any existing value.