feat: return exit code 0 on successful shutdown
The default behavior of Java applications is to return exit code
128 + signal if the application was shut down due to receiving a
signal. This means that SIGINT returns 130 and SIGTERM returns 143.

This change modifies this behavior so exit code 0 is returned if
the shutdown is successful. The old behavior that returns the JVM
default can be restored by starting PGAdapter with the -exit_mode
startup argument.