perf: use multiplexed sessions for read-only operations
Enable multiplexed sessions for all read-only operations. Multiplexed
sessions can handle any number of concurrent read-only operations,
which means that PGAdapter does not need to reserve a session
exclusively for a single read-only operation. This increases the
maximum possible concurrency without the need to increase the
number of sessions in the pool.