feat: support max_commit_delay for COPY
The SQL statement `set spanner.max_commit_delay='20ms'` or similar can be
used to set a maximum commit delay that is allowed for a transaction,
which again can improve throughput for writes. This setting was not used
by PGAdapter for `COPY` statements. This changes ensures that the
`max_commit_delay` that has been set for a connection will also be used
for any `COPY` statement on that connection.