feat: add option for partitioned COPY TO STDOUT
Adds the setting 'spanner.copy_partition_query=true|false' to enable/disable the
use of PartitionQuery for COPY TO STDOUT operations. Using PartitionQuery can
improve the throughput when reading large amounts of data. However, the
additional roundtrips and additional complexity of partitioning a query can be
less efficient for smaller queries. This option allows the user to determine
whether PGAdapter should try to use PartitionQuery or not.

The default is true.

Fixes #2224