feat: support all 'spanner.*' options in pg_settings
Re-writes the pg_settings common table expression to use an array of
jsonb instead of a list of UNION ALL statements. This circumvents the
limitation that a query cannot contain more than 60 UNION ALLs,
which means that we can put a larger value for the limitation on the
number of settings included in the CTE. We still restrict the number
of elements in the CTE to prevent the query text from becoming too
long.