feat: add option for logging gRPC messages (#2230)
Adds an option for logging all gRPC messages that are sent by
PGAdapter to Spanner to make it easier to get insights into
what exactly PGAdapter and/or an application does. This option
should only be used for debugging, and is not recommended for
production use, as it will produce a large amount of log lines.