fix: wait for all connections to close before stopping (#2304)
* fix: wait for all connections to close before stopping

When shutting down PGAdapter using SMART mode, it was theoretically
possible that PGAdapter would terminate before all connections were
closed. This could happen if a new connection came in at exactly the
same time as that the shutdown request was sent. This could cause the
shutdown latch to be created based on the number of connections at that
time, while the new connection would be created and then fail directly
due to connection listener being closed. This again would de-register
the new (failed) connection from the server, which would trigger a count
down of the latch that the SMART shutdown was waiting for.

This is now fixed by simply creating a latch with a value of 1 whenever
SMART shutdown is requested and there is at least one connection open.

Fixes #2300

* fix: create open latch if there are no connections