docs: add sample for using Batch DML with Spring Data JPA
Adds a sample that shows how to use Batch DML in an application that
uses Spring Data JPA / Hibernate to communicate with PGAdapter.

Executing a large number of DML statements as a single DML batch
can significantly reduce the number of round-trips to Spanner, and
thereby the overall latency of a transaction that inserts/updates
a large number of records.

This feature requires version 0.39.0 or higher of PGAdapter.