feat: add session variable for batch_dml_update_count (#2356)
PGAdapter by default returns update count 0 for DML statements that
are executed during a DML batch. The reason for this is that the
DML statement is at that moment only buffered in PGAdapter, and
PGAdapter does not know what the actual update count will be once
it is executed. The PostgreSQL wire-protocol expects the value to
be non-negative, meaning that a negative value as a placeholder is
not possible.
The default update count of 0 can cause problems with frameworks
that expect the update count to be non-zero or a specific value.
Hibernate for example expects each insert to insert exactly one
row, and has a hard-coded check that verifies that the insert
statemnent actually returned an update count of 1. This change
adds a session variable `spanner.dml_batch_update_count` that
determines the update count that PGAdapter returns for DML
statements that are executed/buffered during a DML batch. This
can be used to execute batches of DML statements in for example
Hibernate.
PGAdpater still returns the actual update counts when the DML
batch is executed on Spanner. This information is however not
something that Hibernate and other frameworks use, as it is
not part of the standard JDBC API or other standards.