feat: support SHUTDOWN SQL command (#2388)
* feat: support SHUTDOWN SQL command

Adds support for a custom SQL command to shut down the server:

```
SHUTDOWN [SMART | FAST | IMMEDIATE]
```

This command only shuts down the server if PGAdapter has been started with
the command line argument --allow_shutdown_command. This argument should
only be used when PGAdapter runs in a trusted environment, for example as
a side-car container. The SQL statement can be used to shut down PGAdapter
after finishing a batch job, or when the client application knows that the
cluster / pod / VM / ... it is running in is shutting down.

Fixes #2040

* chore: move shutdown handler to ProxyServer