feat: add option for end-to-end tracing
Add a command line argument and option for enabling end-to-end
tracing when running PGAdapter. This will propagate the current
trace context to the Spanner server, and collect server-side
traces together with the traces that are collected by PGAdapter.

The server-side traces can only be exported to Google Cloud Trace.
Make sure that you have set up an OpenTelemetry instance that
exports traces to Google Cloud Trace to use this feature.