test: fix flaky random results test
The random results test would sometimes fail for dates in the year 100.
This was caused by the getDate(..) method using the local timezone to
calculate the date based on the number of days since the PG epoch.

Instead of using the getDate(..) method, the date should be retrieved
as a LocalDate, as that does not use any timezone information or rely
on any of the weird java.util.Date / java.sql.Date types.

Fixes #2688