fix: ignore unrecognized portal name in Close messages
Ignore it and return a success message if a client tries to close a portal
that no longer exists. Portals are automatically closed when a transaction
ends, and a client might not correctly keep track of this. Just ignore it
if a client tries to close a portal that no longer exists.