perf: use a stream mark to read null-terminated string
Prepared statements and portals can have names in the PG wire-protocol.
These are sent as null-terminated strings. Read these strings by setting
a mark in the stream and then scan for the null-terminator, instead of
always creating a 128-byte buffer and read into that buffer. Using a
stream mark will instead re-use the existing buffer in the existing
input stream, and so reduce the amount of memory that is being reserved
and released for each statement that is received.