fix: ignore TCP-aliveness check connections
Aliveness checks that only open a TCP connection to PGAdapter could cause an EOF
error to be printed by PGAdapter, as these connections close while PGAdapter is
waiting for a PostgreSQL startup message. This makes PGAdapter just ignore EOF
failures at that point of the connection phase, as it is very likely that such a
connection came from a TCP aliveness prober that checks whether PGAdapter is
running or not.