chore: move control message creation to separate class
Move the logic for creating a control message to a separate MessageReader
class. This makes it easier to optimize this specific part of the code for
specific use cases, e.g. introducing a busy-wait for messages in low-QPS
scenarios.

This change only moves the code and adds an option for busy-waiting, but
does not enable it by default in any scenario.