chore: only use original SQL statement with comments
PGAdapter should not use the SQL string that has been stripped for comments
anywhere, as the built-in parser is able to handle comments in the SQL string.
The getSqlWithoutComments() method in the Java client library will eventually
be deprecated, which will remove some additional string parsing and creation
for each statement that is executed.