fix: named prepared statement could return cryptic error message (#3921)
Creating a named prepared statement with a SQL string that would lead to an error,
and that contained at least one query parameter, could lead to the cryptic error
'Statement result cannot be retrieved before flush/sync' to be returned. This would
happen if the prepared statement was executed using at least one untyped query
parameter value, as this would trigger an (auto-)describe of the parameters in the
prepared statement. An error would then be returned from Spanner, as the SQL string
was invalid. This error would however not be returned to the client, and instead no
result would be registered for the statement. That again would lead to the cryptic
error message being returned (which means 'no result found for this statement').